added site model cmds and few refactor
