+++ deepomatic/cli/cmds/site/model/blur.py
@@ -1,5 +1,5 @@
 from deepomatic.cli.cmds.utils import Command
-from ..utils import InferManager, BlurImagePostprocessing
+from deepomatic.cli.lib.inference import InferManager, BlurImagePostprocessing
 from deepomatic.cli.cmds.utils import setup_cmd_line_parsers
