+++ deepomatic/cli/cmds/site/model/draw.py
@@ -1,5 +1,5 @@
 from deepomatic.cli.cmds.utils import Command
-from ..utils import InferManager, DrawImagePostprocessing
+from deepomatic.cli.lib.inference import InferManager, DrawImagePostprocessing
 from deepomatic.cli.cmds.utils import setup_cmd_line_parsers
