+++ deepomatic/cli/cmds/system/__init__.py
@@ -7,14 +7,6 @@ class SystemCommand(Command):
         System command
     """
 
-    class SetupCommand(Command):
-        """
-            Setup the system
-        """
-
-        def run(self, **kwargs):
-            SystemManager().setup()
-
     class CheckCommand(Command):
         """
             Check the system
